requirements.txt
anyio==3.6.1
Arpeggio==1.10.2
asgiref==3.5.2
attrs==21.4.0
beautifulsoup4==4.11.1
boto3==1.24.12
botocore==1.27.12
bs4==0.0.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
decorator==4.4.2
dill==0.3.5.1
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
faunadb==4.2.0
filelock==3.7.1
future==0.18.2
geocoder==1.38.1
greenlet==1.1.2
grpcio==1.46.3
h11==0.13.0
h2==2.6.2
hpack==3.0.0
httptools==0.4.0
huepy==1.2.1
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
iniconfig==1.1.1
instabot==0.117.0
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mock==4.0.3
moviepy==1.0.3
numpy==1.21.6
orjson==3.7.2
packaging==21.3
parver==0.3.1
Pillow==9.1.1
pipenv==2022.6.7
platformdirs==2.5.2
pluggy==1.0.0
proglog==0.1.10
protobuf==4.21.1
psycopg2-binary==2.9.3
pulumi==3.33.1
pulumi-aws==5.8.0
pulumi-cloudflare==4.7.0
py==1.11.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytube==12.1.0
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
ratelim==0.1.6
requests==2.28.0
requests-toolbelt==0.9.1
responses==0.21.0
rich==12.4.4
s3transfer==0.6.0
schedule==1.1.0
semver==2.13.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
